Do not give facter warnings if /etc/snmp is not available
Also fix some stylistic issues

+++ lib/facter/snmpd_pass_progs.rb
@@ -1,28 +1,33 @@
 Facter.add('snmpd_pass_progs') do
-  confine :kernel => 'Linux'
+  confine kernel: 'Linux'
 
   setcode do
     confdir = File.join('/', 'etc', 'snmp')
-    pass_scripts = Dir.entries(confdir).select { |entry| entry.start_with?("pass-") }
 
-    output = {}
+    if Dir.exist?(confdir)
+      pass_scripts = Dir.entries(confdir).select { |entry| entry.start_with?('pass-') }
 
-    pass_scripts.each do |script|
-      path = File.join(confdir, script)
-      oid = nil
-      # Read the second line of the script which should contain
-      # the value for the OID.
-      File.open(path) do |f|
-        f.readline
-        oid = f.readline.split("OID: ")[1].chomp
-        f.close
+      output = {}
+
+      pass_scripts.each do |script|
+        path = File.join(confdir, script)
+        oid = nil
+        # Read the second line of the script which should contain
+        # the value for the OID.
+        File.open(path) do |f|
+          f.readline
+          oid = f.readline.split('OID: ')[1].chomp
+          f.close
+        end
+        output[oid] = path
       end
-      output[oid] = path
-    end
-    if output.empty?
-      nil
+      if output.empty?
+        nil
+      else
+        output
+      end
     else
-      output
+      nil
     end
   end
 end
